Split worksheet implementation
